requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
--extra-index-url https://pypi.ngc.nvidia.com
--trusted-host pypi.ngc.nvidia.com

certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
distro==1.9.0
    # via -r requirements.in
idna==3.7
    # via requests
pyjwt==2.8.0
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
urllib3==2.2.2
    # via requests
